QoL change : looping on snippets, labels and taint lists
Makes things much easier when we add new machines, we no longer have
to add them to the snippets, labels or taints maps one by one in both
the config and monsoon's definition. We now only need to define the
controller or worker as we usually would, add it to the appropriate
node role, such as bsy, bsy_head, bus and so on and done.