snippets/barney : fix btrfs
There was an issue where there could be several btrfs volumes on the
machine, and to recreate our data2 btrfs volume, we would select all
the disks used for btrfs, thinking naively that we would only have one
anyway. However, when installing flatcar on disk, a btrfs volume named
"OEM" is created. This commit fixes that by only selecting the data2
btrfs volume.